llm/receiptScan/requirements.txt
flask==3.0.0
werkzeug==3.0.1
pytesseract==0.3.10
Pillow==10.1.0
langchain-google-genai==0.0.11
google-generativeai==0.4.1
PyMuPDF==1.23.8
numpy==1.26.3
opencv-python-headless==4.9.0.80
